refactor : 폴더 구조 및 상태 위치 수정
